STEAM-533
Added the Prediction Service README as an appendix to the User Guide.
Linked to this appendix from within the Making Predictions section.
Also updated the text within the Making Predictions section to
highlight the purpose of the Prediction Service tool.
These changes required a minor change to the CLI appendix, which is no
longer the only appendix.
Driveby fixes: Changed the background color of the help system from H2O
yellow to Steam orange. Also, in the installation instructions, moved
the step for starting the jetty server to BEFORE starting Steam.
Hopefully there is now less chance that this step is overlooked.