Changed file format from markdown to rst
The readme in the prediction-builder-service folder is now an rst file
instead of an md file. This is so that it is compatible with the
Sphinx-generated HTML User Guide. RST is still simple, plain text, so
anyone contributing should not encounter any issues with this change.
Driveby fixes: Removed unused/older images that still resided in the
prediction-service-builder folder. Fixed an indentation issue with a
note in the Installation section.