STEAM-578: Add support Deep Water in Steam
In order to simplify adding Deep Water (and future different models),
the compilation step from Steam was refactored. It now uses a
ModelAsset interface which is satisfied by having a custom AttachFiles
function. This creates the buffer that satisfies the fields for the
Prediction Service Builder to compile the model.

Currently Deep Water can only build a MOJO, only the ImportModelMojo
function downloads the Deep Water dependency.